requirements.txt
streamlit
sentence-transformers
pdfplumber
python-docx
pytesseract
Pillow
docx2txt
pymupdf
